+++ projects/demo-playwright/tests/kit/input-month/input-month.pw.spec.ts
@@ -25,7 +25,10 @@ describe('InputMonth', () => {
 
         describe('dropdown', () => {
             test('opens on click for NOT readonly input', async ({page}) => {
-                await tuiGoto(page, `${DemoRoute.InputMonth}/API?readOnly=false`);
+                await tuiGoto(
+                    page,
+                    `${DemoRoute.InputMonth}/API?readonly=false&readOnly=false`, // TODO: delete `readOnly` param in next PR
+                );
 
                 await expect(inputMonth.calendar).not.toBeAttached();
                 await inputMonth.textfield.click();
@@ -33,7 +36,10 @@ describe('InputMonth', () => {
             });
 
             test('does NOT open on click for readonly input', async ({page}) => {
-                await tuiGoto(page, `${DemoRoute.InputMonth}/API?readOnly=true`);
+                await tuiGoto(
+                    page,
+                    `${DemoRoute.InputMonth}/API?readonly=true&readOnly=true`, // TODO: delete `readOnly` param in next PR
+                );
 
                 await expect(inputMonth.calendar).not.toBeAttached();
                 await inputMonth.textfield.click();
@@ -85,12 +91,12 @@ describe('InputMonth', () => {
                 await expect(inputMonth.calendar).toBeAttached();
             });
 
-            test('opens dropdown on click on calendar icon (disabled=false&readOnly=false)', async ({
+            test('opens dropdown on click on calendar icon (disabled=false&readonly=false)', async ({
                 page,
             }) => {
                 await tuiGoto(
                     page,
-                    `${DemoRoute.InputMonth}/API?disabled=false&readOnly=false`,
+                    `${DemoRoute.InputMonth}/API?disabled=false&readonly=false&readOnly=false`, // TODO: delete `readOnly` param in next PR
                 );
 
                 await expect(inputMonth.calendar).not.toBeAttached();
@@ -104,7 +110,7 @@ describe('InputMonth', () => {
             }) => {
                 await tuiGoto(
                     page,
-                    `${DemoRoute.InputMonth}/API?disabled=true&readOnly=false`,
+                    `${DemoRoute.InputMonth}/API?disabled=true&readonly=false&readOnly=false`, // TODO: delete `readOnly` param in next PR
                 );
 
                 await expect(inputMonth.calendar).not.toBeAttached();
@@ -118,7 +124,7 @@ describe('InputMonth', () => {
             }) => {
                 await tuiGoto(
                     page,
-                    `${DemoRoute.InputMonth}/API?disabled=false&readOnly=true`,
+                    `${DemoRoute.InputMonth}/API?disabled=false&readonly=true&readOnly=true`, // TODO: delete `readOnly` param in next PR
                 );
 
                 await expect(inputMonth.calendar).not.toBeAttached();