+++ projects/demo/src/pages/components/input-time/index.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 import {FormControl, ReactiveFormsModule} from '@angular/forms';
+import {TuiDocAppearance} from '@demo/components/appearance';
 import {TuiDocControl} from '@demo/components/control';
-import {TuiDocInput} from '@demo/components/input';
 import {TuiDocTextfield} from '@demo/components/textfield';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {DemoRoute} from '@demo/routes';
@@ -15,8 +15,8 @@ import {TuiInputTime} from '@taiga-ui/kit';
     imports: [
         ReactiveFormsModule,
         TuiDemo,
+        TuiDocAppearance,
         TuiDocControl,
-        TuiDocInput,
         TuiDocTextfield,
         TuiInput,
         TuiInputTime,