+++ projects/experimental/components/calendar/calendar.ts
@@ -1,0 +1,4 @@
+import {TuiCalendarComponent} from './calendar.component';
+import {TuiCalendarMultiComponent} from './calendar-multi.component';
+
+export const TuiCalendar = [TuiCalendarComponent, TuiCalendarMultiComponent] as const;