+++ projects/experimental/components/calendar/index.ts
@@ -1,0 +1,5 @@
+export * from './calendar';
+export * from './calendar.component';
+export * from './calendar-header.component';
+export * from './calendar-multi.component';
+export * from './date-picker';