+++ projects/kit/components/calendar-range/test/calendar-range.component.spec.ts
@@ -34,6 +34,7 @@ describe('rangeCalendarComponent', () => {
                 [maxLength]="maxLength"
                 [min]="min"
                 [minLength]="minLength"
+                [(month)]="month"
                 [(value)]="value"
             />
         `,
@@ -50,6 +51,7 @@ describe('rangeCalendarComponent', () => {
         public value: TuiDayRange | null = null;
         public defaultViewedMonth = TuiMonth.currentLocal();
         public markerHandler: TuiMarkerHandler | null = null;
+        public month = TuiMonth.currentLocal();
     }
 
     let fixture: ComponentFixture<Test>;
@@ -383,6 +385,38 @@ describe('rangeCalendarComponent', () => {
         });
     });
 
+    describe('month change', () => {
+        beforeEach(() => {
+            testComponent.items = [];
+            fixture.detectChanges();
+        });
+
+        it('append month by forward chevron', () => {
+            const month = TuiMonth.currentLocal();
+
+            clickNextMonth();
+
+            expect(testComponent.month.monthSame(month.append({month: 1}))).toBe(true);
+        });
+
+        it('reduce month by back chevron', () => {
+            const month = TuiMonth.currentLocal();
+
+            clickPreviousMonth();
+
+            expect(testComponent.month.monthSame(month.append({month: -1}))).toBe(true);
+        });
+
+        it('use month from value', () => {
+            const day = TuiDay.currentLocal().append({month: 5});
+
+            testComponent.value = new TuiDayRange(day, day);
+            fixture.detectChanges();
+
+            expect(testComponent.month.monthSame(day)).toBe(true);
+        });
+    });
+
     function getCalendar(): DebugElement | null {
         return pageObject.getByAutomationId('tui-calendar-range__calendar');
     }
@@ -394,5 +428,19 @@ describe('rangeCalendarComponent', () => {
 
     function getItems(): DebugElement[] {
         return pageObject.getAllByAutomationId('tui-calendar-range__menu__item');
+    }
+
+    function clickPreviousMonth(): void {
+        pageObject
+            .getAllByAutomationId('tui-spin-button__left')[0]!
+            .nativeElement.click();
+        fixture.detectChanges();
+    }
+
+    function clickNextMonth(): void {
+        pageObject
+            .getAllByAutomationId('tui-spin-button__right')[1]!
+            .nativeElement.click();
+        fixture.detectChanges();
     }
 });